Flask/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Flask==2.0.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
